test(tests): update test cases for various mathematical functions and expressions
- Refactor test cases to improve clarity and correctness.
- Adjusted expressions in tests to ensure accurate evaluations.
- Added new tests for edge cases and specific scenarios.

+++ ConstExpr.Tests/Tests/Rewriter/VisitCastExpressionTests.cs
@@ -8,17 +8,17 @@ public class VisitCastExpressionTests : BaseTest<Func<double, int, int, (int, do
 {
 	public override string TestMethod => GetString((x, y, z) =>
 	{
-		var a = (int)x;
-		var b = (double)y;
-		var c = (char)z;
-		var d = (int)3.14;
+		var a = (int) x;
+		var b = (double) y;
+		var c = (char) z;
+		var d = (int) 3.14;
 
 		return (a, b, c, d);
 	});
 
 	public override IEnumerable<KeyValuePair<string?, object?[]>> TestCases =>
 	[
-		Create((x, y, z) => ((int)x, (double)y, (char)z, 3)),
+		Create((x, y, z) => ((int) x, y, (char) z, 3)),
 		Create((_, _, _) => (3, 42.0, 'A', 3), [ 3.14, 42, 65 ]),
 		Create((_, _, _) => (10, 100.0, 'Z', 3), [ 10.5, 100, 90 ]),
 		Create((_, _, _) => (-5, -10.0, ' ', 3), [ -5.8, -10, 32 ])
